+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2025-11-25
+
+- Always index resources in a separate process
+
+  Indexing resources in the same process as the one modifying resources can
+  cause a lot of different issues: performance, memory usage, conflicting use
+  of the Doctrine entity manager causing errors or data duplication, ...
+
+  Doing it in a separate process ensure that we do not disturb the main process.
+
+  There is a new background job (`Search\Job\Sync`) that is run periodically
+  when there are resources that need to be indexed.
+  Alternatively, there is also a new script (`bin/sync`) that can be executed
+  periodically (using cron for instance). It does the same thing as the new
+  background job.
+
+- Show the latest indexation date and time on admin resource details page
+
 ## [0.17.5] - 2025-04-16
 
 - Fix missing TestAdapter methods (causing tests failure)